bugfixes in loadDL()
git-svn-id: https://svn.code.sf.net/p/socnetv/code@166 fc2d5e9b-b3ce-4657-a66d-b27cdf764933

+++ trunk/src/parser.h
@@ -55,7 +55,7 @@ public:
 	void dotProperties(QString str, int &nValue, QString &label, QString &shape, QString &color, QString &fontName, QString &fontColor );
 signals:
 	void createNode(int,int,QString, QString, QString, QPointF, QString, bool);
-	void createEdge (int, int, int, QString, bool, bool, bool);
+	void createEdge (int, int, float, QString, bool, bool, bool);
 	void fileType(int, QString, int, int);
 	void removeDummyNode (int);
 protected:
